refactor: provider and extractor 분리
