+++ src/main/java/pretzel/dreamketcherbe/auth/jwt/JwtExtractor.java
@@ -1,0 +1,39 @@
+package pretzel.dreamketcherbe.auth.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.JwtParser;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import pretzel.dreamketcherbe.auth.dto.AuthPayload;
+import pretzel.dreamketcherbe.member.entity.Role;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+
+@Component
+public class JwtExtractor implements TokenExtractor {
+
+    private static final String MEMBER_ID = "memberId";
+    private static final String ROLE = "role";
+
+    private final JwtParser jwtParser;
+
+    public JwtExtractor(@Value("${jwt.secret-key}") String secretKeyString){
+        SecretKey secretKey = Keys.hmacShaKeyFor(secretKeyString.getBytes(StandardCharsets.UTF_8));
+        this.jwtParser = Jwts.parserBuilder()
+                                .setSigningKey(secretKey)
+                                .build();
+    }
+
+    @Override
+    public AuthPayload extract(String token){
+        Claims claims = jwtParser.parseClaimsJws(token).getBody();
+        Long memberId = claims.get(MEMBER_ID, Long.class);
+        String role = claims.get(ROLE, String.class);
+        return new AuthPayload(memberId, Role.of(role));
+    }
+}
