+++ src/main/java/pretzel/dreamketcherbe/auth/jwt/TokenExtractor.java
@@ -1,9 +1,8 @@
 package pretzel.dreamketcherbe.auth.jwt;
 
-
 import pretzel.dreamketcherbe.auth.dto.AuthPayload;
 
-public interface TokenProvider {
+public interface TokenExtractor {
 
-    String generated(AuthPayload authPayload);
+    AuthPayload extract(String token);
 }
